Added hassembler to interface BEAST directly
It is debatable if FastBEAST should include BEAST really.

Given that it is called FastBEAST, I'd argued now that such a
directly link is acceptable.

In future, one could consider splitting up FastBEAST in more packages
and thus remove this hard link to BEAST.

Hence, some of the machinery could be used with other packages, too.