Refined boxtree vs kmeans performance unit tests
TODO: for h=0.008 in the surface performance test,
the performance of kmeans is too slow to serve as a unit test
in practice.

Either deactivate or use parallelizationTODO: for h=0.008 in the surface
performance test,
the performance of kmeans is too slow to serve as a unit test
in practice.

Either deactivate or use parallelization..