Modification of hassemble parameters and unit tests
The parameters of the hassemble functions were not structured in clean way.
I combined all parameters of the compression in one Type. The current
state of the package only provides a ACA compressor with different
pivoting strategies.
I fixed some unit tests and change the function calls to work with the
new parameter structure

ToDo:
- The fmmassamble parameters should be structured likewise the hassemble.