Modifiacations towards a flexible ACA
The ACA did not allow different convergence criteria yet.
The default criterion is still the stadard one following Zhao.
Additionally I included a random sampling and a combined criterion.
Further criteria can be included following the scheme given in the
convergence.jl script.

Beside these modifications I included in the H-Matrix part a check for
ineffective approximations. If the ACA uses to many rows and columns, the
H-Matrix computes the full matrix block and neglects the approximation.