Rework of the ACA algorithm
I reworked the ACA towards a more flexible implementation.

I renamed the 'TrueFillDistance' to 'FillDistance' and introduced an
abstract type 'FD' that concludes all filldistance based strategies.
This allows to reduce code repetition.
Additionally I included the 'MRFPivoting' strategy.

As convergence criteria the 'Standard' criterion, a 'Random Sampling',
and a 'Combined' criterion are available.

The modifications in the 'test_beast.jl' shows that the compression
of matrix blocks becomes more efficient due to better handling of
special cases such as zero rows or similar.