Basic documentation for FastBEAST
I wrote a basic documentation for FastBEAST.
This is work and progress and not finished.

ToDo:
Fix the unit tests and set up UI and enforce style guide,
the we can add a contributing file.
I think we might introduce a FMM type with a FMM function analog to the HMatrix
function, then we can add the API to H-matrix and the FMM.
Include ClusterTrees for the box tree, then we can update the example
for clustering.